+++ components/dash-core-components/src/types.ts
@@ -416,9 +416,9 @@ export interface DropdownProps extends BaseComponentProps<DropdownProps> {
     style?: React.CSSProperties;
 
     /**
-     * Translations for customizing text contained within this component.
+     * Text for customizing the labels rendered by this component.
      */
-    localizations?: {
+    labels?: {
         select_all?: string;
         deselect_all?: string;
         selected_count?: string;