+++ components/dash-core-components/tests/integration/dropdown/test_localization.py
@@ -11,7 +11,7 @@ def test_ddlo001_translations(dash_duo):
                 id="dropdown",
                 options=[1, 2, 3],
                 multi=True,
-                localizations={
+                labels={
                     "select_all": "Sélectionner tout",
                     "deselect_all": "Désélectionner tout",
                     "selected_count": "{num_selected} sélections",
@@ -28,32 +28,27 @@ def test_ddlo001_translations(dash_duo):
 
     dash_duo.find_element("#dropdown").click()
     dash_duo.wait_for_contains_text(
-        "#dropdown .dash-dropdown-action-button:first-child", "Sélectionner tout"
+        ".dash-dropdown-action-button:first-child", "Sélectionner tout"
     )
     dash_duo.wait_for_contains_text(
-        "#dropdown .dash-dropdown-action-button:last-child", "Désélectionner tout"
+        ".dash-dropdown-action-button:last-child", "Désélectionner tout"
     )
 
-    search_input = dash_duo.find_element("#dropdown .dash-dropdown-search")
-    assert search_input.accessible_name == "Rechercher"
-
-    search_input.send_keys(1)
     assert (
-        dash_duo.find_element("#dropdown .dash-dropdown-clear").accessible_name
-        == "Annuler"
+        dash_duo.find_element(".dash-dropdown-search").accessible_name == "Rechercher"
     )
 
-    dash_duo.find_element("#dropdown .dash-dropdown-action-button:first-child").click()
+    dash_duo.find_element(".dash-dropdown-search").send_keys(1)
+    assert dash_duo.find_element(".dash-dropdown-clear").accessible_name == "Annuler"
 
-    search_input.send_keys(9)
-    assert (
-        dash_duo.find_element("#dropdown .dash-dropdown-option").text
-        == "Aucun d'options"
-    )
+    dash_duo.find_element(".dash-dropdown-action-button:first-child").click()
+
+    dash_duo.find_element(".dash-dropdown-search").send_keys(9)
+    assert dash_duo.find_element(".dash-dropdown-option").text == "Aucun d'options"
 
     assert (
         dash_duo.find_element(
-            "#dropdown .dash-dropdown-trigger .dash-dropdown-clear"
+            ".dash-dropdown-trigger .dash-dropdown-clear"
         ).accessible_name
         == "Effacer les sélections"
     )
@@ -69,7 +64,7 @@ def test_ddlo002_partial_translations(dash_duo):
                 id="dropdown",
                 options=[1, 2, 3],
                 multi=True,
-                localizations={
+                labels={
                     "search": "Lookup",
                 },
             ),
@@ -80,32 +75,27 @@ def test_ddlo002_partial_translations(dash_duo):
 
     dash_duo.find_element("#dropdown").click()
     dash_duo.wait_for_contains_text(
-        "#dropdown .dash-dropdown-action-button:first-child", "Select All"
+        ".dash-dropdown-action-button:first-child", "Select All"
     )
     dash_duo.wait_for_contains_text(
-        "#dropdown .dash-dropdown-action-button:last-child", "Deselect All"
+        ".dash-dropdown-action-button:last-child", "Deselect All"
     )
 
-    search_input = dash_duo.find_element("#dropdown .dash-dropdown-search")
-    assert search_input.accessible_name == "Lookup"
+    assert dash_duo.find_element(".dash-dropdown-search").accessible_name == "Lookup"
 
-    search_input.send_keys(1)
+    dash_duo.find_element(".dash-dropdown-search").send_keys(1)
     assert (
-        dash_duo.find_element("#dropdown .dash-dropdown-clear").accessible_name
-        == "Clear search"
+        dash_duo.find_element(".dash-dropdown-clear").accessible_name == "Clear search"
     )
 
-    dash_duo.find_element("#dropdown .dash-dropdown-action-button:first-child").click()
+    dash_duo.find_element(".dash-dropdown-action-button:first-child").click()
 
-    search_input.send_keys(9)
-    assert (
-        dash_duo.find_element("#dropdown .dash-dropdown-option").text
-        == "No options found"
-    )
+    dash_duo.find_element(".dash-dropdown-search").send_keys(9)
+    assert dash_duo.find_element(".dash-dropdown-option").text == "No options found"
 
     assert (
         dash_duo.find_element(
-            "#dropdown .dash-dropdown-trigger .dash-dropdown-clear"
+            ".dash-dropdown-trigger .dash-dropdown-clear"
         ).accessible_name
         == "Clear selection"
     )