+++ components/dash-core-components/tests/integration/misc/test_persistence.py
@@ -131,18 +131,18 @@ def test_msps001_basic_persistence(dash_dcc):
     dash_dcc.select_date_single("datepickersingle", day="20")
 
     dash_dcc.find_element("#dropdownsingle").click()
-    dash_dcc.find_element("#dropdownsingle .dash-dropdown-search").send_keys(
+    dash_dcc.find_element(".dash-dropdown-content .dash-dropdown-search").send_keys(
         "one" + Keys.ENTER
     )
     sleep(0.2)
-    dash_dcc.find_element("#dropdownsingle .dash-dropdown-option").click()
+    dash_dcc.find_element(".dash-dropdown-content .dash-dropdown-option").click()
 
     dash_dcc.find_element("#dropdownmulti").click()
-    dash_dcc.find_element("#dropdownmulti .dash-dropdown-search").send_keys(
+    dash_dcc.find_element(".dash-dropdown-content .dash-dropdown-search").send_keys(
         "six" + Keys.ENTER
     )
     sleep(0.2)
-    dash_dcc.find_element("#dropdownmulti .dash-dropdown-option").click()
+    dash_dcc.find_element(".dash-dropdown-content .dash-dropdown-option").click()
 
     dash_dcc.find_element("#input").send_keys(" maybe")
 
@@ -168,7 +168,7 @@ def test_msps001_basic_persistence(dash_dcc):
         ["4️⃣", "6️⃣"],
         "yes maybe",
         "r",
-        [5, 9],
+        [5, 8],
         22,
         "C",
         "knock knock\nwho's there?",