+++ components/dash-core-components/tests/integration/misc/test_platter.py
@@ -6,7 +6,6 @@ from selenium.common.exceptions import WebDriverException
 
 
 def test_mspl001_dcc_components_platter(platter_app, dash_dcc):
-
     dash_dcc.start_server(platter_app)
 
     dash_dcc.wait_for_element("#waitfor")
@@ -17,7 +16,7 @@ def test_mspl001_dcc_components_platter(platter_app, dash_dcc):
     dash_dcc.percy_snapshot("gallery")
 
     dash_dcc.find_element("#dropdown").click()
-    dash_dcc.find_element("#dropdown .dash-dropdown-search").send_keys("北")
+    dash_dcc.find_element(".dash-dropdown-content .dash-dropdown-search").send_keys("北")
     dash_dcc.percy_snapshot("gallery - chinese character")
 
     text_input = dash_dcc.find_element("#textinput")