+++ dash/testing/browser.py
@@ -432,7 +432,7 @@ class Browser(DashPageMixin):
         dropdown = self._get_element(elem_or_selector)
         dropdown.click()
 
-        menu = dropdown.find_element(By.CSS_SELECTOR, ".dash-dropdown-options")
+        menu = self._get_element(".dash-dropdown-options")
         logger.debug("the available options are %s", "|".join(menu.text.split("\n")))
 
         options = menu.find_elements(By.CSS_SELECTOR, ".dash-dropdown-option")